+++ src/dto-interfaces/user.dto.ts
@@ -2,5 +2,6 @@ export interface DtoUser {
   condition?     : string
   documentNumber?: string
   documentType?  : string
+  gender?        : string
   id?            : string
 }