+++ src/utils/index.ts
@@ -1,5 +1,5 @@
 import docs from './docs.json'
-import { mail } from './mailer'
+import { deliverPassword, notifyProcuratorRegistered } from './mailer'
 import { MFE, MFME } from './messages'
 import { PATA } from './constants'
 import { decryptMessage, encryptMessage, generatePassword } from './security'
@@ -12,14 +12,15 @@ import {
 } from './jwt.helper'
 
 export {
+  deliverPassword,
   docs,
-  mail,
   MFE,
   MFME,
   PATA,
   decryptMessage,
   encryptMessage,
   generatePassword,
+  notifyProcuratorRegistered,
   response,
   signAccessToken,
   signRefreshToken,